Move controllers from pro to worker/server
