--- packages/pro/src/api/controllers/global/index.ts
@@ -1,2 +1,0 @@
-export * as groups from "./groups"
-export * as environmentVariables from "./environmentVariables"
