+++ packages/pro/src/index.ts
@@ -1,8 +1,8 @@
 export * from "./types"
 export * from "./sdk"
 export * as sdk from "./sdk"
+export * as db from "./db"
 export * as constants from "./constants"
 export * as middleware from "./middleware"
-export * as api from "./api"
 export * as mappers from "./mappers"
 export * as ai from "./ai"
