+++ packages/server/src/api/routes/index.ts
@@ -1,6 +1,7 @@
-import { api as pro } from "@budibase/pro"
 import Router from "@koa/router"
 import { endpointGroupList } from "./endpointGroups"
+import appBackupRoutes from "./apps/backups"
+import environmentVariableRoutes from "./environmentVariables"
 
 // just need to import routes, they'll include themselves in the
 // various groups they need to be included in
@@ -45,9 +46,6 @@ import "./workspaceFavourites"
 export { default as assetRoutes } from "./assets"
 export { default as publicRoutes } from "./public"
 export { default as staticRoutes } from "./static"
-
-const appBackupRoutes = pro.appBackups
-const environmentVariableRoutes = pro.environmentVariables
 const endpoints = endpointGroupList.listAllEndpoints()
 
 const appRoutes = new Router()
