+++ packages/worker/src/api/routes/index.ts
@@ -1,4 +1,3 @@
-import { api as pro } from "@budibase/pro"
 import Router from "@koa/router"
 import { endpointGroupList } from "./endpointGroups"
 
@@ -18,6 +17,9 @@ import "./system/logs"
 import "./system/restore"
 import "./system/status"
 import "./system/tenants"
+import auditLogsRoutes from "./global/auditLogs"
+import groupRoutes from "./global/groups"
+import scimRoutes from "./global/scim"
 
 const endpointGroupsRouter = new Router()
 for (let endpoint of endpointGroupList.listAllEndpoints()) {
@@ -26,7 +28,7 @@ for (let endpoint of endpointGroupList.listAllEndpoints()) {
 
 export const routes: Router[] = [
   endpointGroupsRouter,
-  pro.groups,
-  pro.auditLogs,
-  pro.scim,
+  auditLogsRoutes,
+  groupRoutes,
+  scimRoutes,
 ]
